requirements.txt
# IDAES Process Simulation Framework
idaes-pse==2.8.0

# AutoGen Multi-Agent Framework
autogen-agentchat==0.5.1
autogen-core==0.5.1
autogen-ext==0.5.1

# Google Gemini and LLM Dependencies
google-generativeai==0.3.2
tiktoken==0.9.0

# Data Processing
pandas==2.2.3
pyyaml==6.0.2
